requirements.txt
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
pygame>=2.5.1
gitpython>=3.1.30
Adafruit-Blinka>=8.20.1
Adafruit-PlatformDetect>=3.49.0
Adafruit-PureIO>=1.1.11
adafruit-circuitpython-busdevice>=5.2.6
adafruit-circuitpython-fingerprint>=2.2.12
adafruit-circuitpython-requests>=2.0.1
adafruit-circuitpython-typing>=1.9.4
# dlib>=19.24.2
imutils>=0.5.4
numpy>=1.25.2
opencv-contrib-python>=4.7.0.72

matplotlib>=3.5.1
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.27.0
scipy>=1.10.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.15.2
tqdm>=4.64.0
ultralytics>=8.0.173
wheel>=0.40.0
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting --------------------------------------------------------------------
pandas>=1.4.2
seaborn>=0.11.0

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.10.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Deploy ----------------------------------------------------------------------
setuptools>=65.5.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
pyserial~=3.5
